Changed makelink.sh to not create dotfiles_old dir unless necessary

+++ makelinks.sh
@@ -12,11 +12,6 @@ files="bashrc vimrc" # vim zshrc oh-my-zsh"    # list of files/folders to symlin
 
 ##########
 
-# create dotfiles_old in homedir
-echo "Creating $olddir for backup of any existing dotfiles in ~"
-mkdir -p $olddir
-echo "...done"
-
 # change to the dotfiles directory
 echo "Changing to the $dir directory"
 cd $dir
@@ -24,8 +19,17 @@ echo "...done"
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks 
 for file in $files; do
-	echo "Moving any existing dotfiles from ~ to $olddir"
-	mv ~/.$file ~/dotfiles_old/
+    if [ -f ~/.$file ] && [ ! -d $olddir ];
+    then
+        echo "Creating $olddir for backup of any existing dotfiles in ~"
+        mkdir -p $olddir
+        echo "Moving any existing dotfiles from ~ to $olddir"
+	    mv ~/.$file ~/dotfiles_old/
+    
+    elif [ -f ~/.$file ];
+    then        
+    	mv ~/.$file ~/dotfiles_old/
+    fi
 	echo "Creating symlink to $file in home directory."
 	ln -s $dir/$file ~/.$file
 done
